Annotation Refactoring - Remove duplicated fetchAnnotations function
and other smaller reorganisations
