+++ src/sidebar-overlay/utils.ts
@@ -63,3 +63,5 @@ export const maxPossibleTags = (tags: string[]) => {
     }
     return tagsAllowed
 }
+// FIXME (ch - annotations): replace with typed version
+export const toggleSidebarOverlay = remoteFunction('toggleSidebarOverlay')
